legacy/rag-prototype-main/requirements.txt
streamlit==1.37.0
streamlit-feedback==0.1.3
langchain==0.2.11
langchain_pinecone==0.1.3
langchain_openai==0.1.19
langchain-community==0.2.10
langchain-huggingface==0.0.3 # dependency includes torch, transformers, nvidia_cuda, nvidia_cudnn and etc.
python-dotenv==1.0.1
ragas==0.1.13
datasets==2.20.0
pandas==2.2.2
bitsandbytes==0.43.3
accelerate==0.33.0
